Automated check for Maven dependency updates
The launch config will invoke Maven on the target platform module,
activating a special profile. In that profile, the target files are
parsed for maven coordinates, and those coordinates are injected into
the Maven model of the build, so they can take part in a standard Maven
version update check.

Disallow all kinds of non release versions via property.

Change-Id: I5d9f422c83f5426db465196ba41fae285d3495da