EGitCredentialsProvider: support asking for a password only
Don't try to use the LoginDialog if we only have a password
CredentialItem. The dialog would not return any credentials if
no user name was entered. Instead use the general dialog.

JGit may ask for a password only if a PKCS11 provider is used.
Normally, it will route such prompts through the KeyPasswordProvider
interface, which can handle this. But if the PKCS11 KeyStore
implementation asks for more than a single password or wants additional
information to be displayed, we might end up in this case.

Change-Id: I1d17aeb279609fdd1737b30de6b1faf41bfe7287