History view: dynamic label for "Push Commit..."
This command uses a "Push Branch" dialog if there is a local branch
pointing to the selected commit, and a "Push Commit" dialog otherwise.

The "Push Branch" is sufficiently different from the "Push Commit"
dialog; in particular it also offers to set up a remote tracking branch.
The menu label "Push Commit..." is then a little bit misleading.

Change the command handler to update its label to "Push Branch..." if
there is a local branch pointing to the selected commit.

Bug: 582226
Change-Id: If1de2024c12818cadd69e00a09eacd189c7076f9
Signed-off-by: Thomas Wolf <twolf@apache.org>