Staging View: only use IFile if project is managed by EGit
Ensure that we only use an IContainer for the working tree root if the
container is in a project shared with git. StagingFolderEntry already
did so for determining its IContainer, if any, but for the root, we
inadvertently used non-shared containers, too, which could lead to
problems as in the reported bug.

Bug: 582517
Change-Id: I186eaf3f65d226aa8ca4f38264c2cb9f1dac067d
Signed-off-by: Thomas Wolf <twolf@apache.org>