Revert "EGitCredentialsProvider: support asking for a password only"
This reverts commit e23d0a648a9e7daca4ddff5538aca686c743ca8d. It
bypassed the Eclipse secure storage, which is highly inconvenient
in the case of SSH password authentication.

The change originally was done to support a fallback option in the
PKCS11 provider asking for a PIN to avoid that it asked for a user name
(which isn't needed, and ignored), too. But the PKCS11 code normally
will use the KeyPasswordProvider, which works fine and is also
integrated with the Eclipse secure storage.

SSH password authentication also will ask only for a password, but the
URIish will contain the user name, which EGit will use to pre-fill the
user-name dialog field (and make it read-only).

So in short: the change was misguided and not needed and broke secure
storage integration for SSH password authentication. Reverting it
restores this functionality, and PIN prompts from the PKCS11 provider
also still work.

Change-Id: I47238047490ca26c30c9cc03cf3979a5ae8849bc