Fix jgit lock exception in egit staging view
If a git repository is not linked to a Eclipse project the job locking
mechanism does not work in if files are added and removed to the index
at the same time.
The lock relies on Eclipse job scheduling rules.
Enhance the current implementation for the case that a git repository is
not present as Eclipse project.
Add 2 new test cases: a) git repository is also a Eclipse project, b)
git repository is not a Eclipse project

Change-Id: Ia138bc3e6e5f8fef4d4884869db5cd69d5b6195a