Replace (partly unused) dependency to org.eclipse.osgi.services
The bundle 'o.e.osgi.services' is deprecated for removal as part of
https://github.com/eclipse-equinox/equinox/issues/18
and already has been replaced to only re-export the 'official' OSGi
artifacts published to Maven-Central.

It should be replaced by imports of the desired org.osgi.service.*
packages so that the runtime can choose a suitable provider
independently of the bundle name.