Manage the Paste command when the clipboard contains a git URL
Open the clone wizard if the currently active part is an IViewPart
and the clipboard contains a valid git URL, unless the current focus
control is a Text or StyledText. Don't do so if the currently active
part is an editor part to avoid interfering with editing. Limiting
this feature to ViewParts also disables it in dialogs.

Bug: egit-48
Change-Id: Ic30cf3155d96faf2b460cdc8e681826a8e3ce3fd
Signed-off-by: Olivier Prouvost <olivier.prouvost@opcoach.com>