[releng] Fix settings for Java 17
JGit requires Java 17 as of release 7.0. So does EGit, but the Eclipse
JDT settings still were for Java 11.

Also, set egit-4.28 as default target platform. Eclipse 4.28 (2023-06)
was the first Eclipse version that required Java 17. For building we
should use that target. Also, Eclipse 2023-06 should become the new
baseline for EGit.

The earlier target platform are still kept; they may come in handy if
one wants to work on a 6.x branch.

Change-Id: Ic4fbcc4ec138dc9bfa78c5d7e97544edfc0cd35e
Signed-off-by: Thomas Wolf <twolf@apache.org>