Minor UI fix: show only first line of regex error message
Pattern.compile() may produce an exception message with several lines,
including the pattern text and a ^ on a line below to indicate the
error position.

This looks bad when displayed in an SWT Label, and it's also not
necessary because the message will mention the index, and the user
sees the regex he or she is typing anyway.

Change-Id: Ic6871df0f741b79940a59b139f12ae66c14345ec
Signed-off-by: Thomas Wolf <twolf@apache.org>