Signing: support smimesign for x509 signatures
smimesign is a lightweight alternative to gpgsm for x509 signatures. It
mimics the interface of gpgsm, at least for the bits used in C git.

Make canLocateSigningKey() a no-op for x509: the user must have
explicitly configured something (at least gpg.format = x509), so we may
safely assume that the user intended to be able to sign, and if then no
signing key is available, it's simply a user error and the user will
have to fix the config.

When invoking the program, do pass only the same options as C git does.
The extra options "--batch --no-tty" were a belt-and-suspenders approach
to ensure that the program would prompt for key passphrases in a dialog,
not in some terminal. It appears they're not needed. Likewise, the
arguments "--output -" is superfluous as the output goes to stdout
anyway. smimesign supports only "-bsau <keySpec>" and "--status-fd 2".

In the verifier refactor the timestamp parsing to ensure we can handle
either GPG format in all cases, and extract the verifier name from the
first line of the program's stderr. (gpgsm produces human-readable
output on stderr, all lines prefixed by "gpgsm: ". smimesign mimics
that, but uses "smimesign: " as prefix.)

Note that smimesign does not produce a "VALIDSIG" status line, so we do
not get the signature creation timestamp.

Bug: egit-59
Change-Id: I80e920b46158d6e1da50dd56df810ccbe52e1c96