Signing: locate the program on $PATH
With a git config "gpg.program = gpg" we'd just try to execute "gpg",
which would fail.

Trying to run the executable via FS.runInShell() to let the shell locate
the program on $PATH didn't work for me on Mac OS even though the shell
was invoked as "sh -l". It also has the problem that some arguments
would need to be escaped, which depends on the shell (Unix sh vs.
Windows cmd.exe).

Instead search $PATH explicitly to figure out the full path of the
program to run.

Change-Id: I0cd40147efc06f2dc607ec859debd58067c8ec0a