EGitGpgConfig: restore override functionality
The preference setting for the external gpg-compatible program is
intended to be an *override*, not a *fallback*. The intention was
that if the program defined in the git config didn't work from within
EGit/Eclipse the user could define an alternate executable.

Restore this functionality. Commit b01eccd78 had inadvertently changed
this from being an override to a fallback.

Change-Id: Ica2621336cb66b46dc9efd339e01522ded455fda