Import Team Project Set can create local branch
Currently the configured branch of a project in a Team Project Set
is not used, when the repository of the project is already
checked out and the configured branch has not yet a local branch.
The changes in this commit enhances the import of a Team Project Set
to create a new local branch if there is none for the branch
configured in the Team Project Set.

Change-Id: I2479f8dd31d33b7f55ecbc0972511a3bda233a52
Signed-off-by: Dirk Baumann <dbaumann@dxc.com>