Staging View: fix display for case collisions
On a case-insensitive filesystem, when one renames a file to a case
variant (such as "a.txt" to "A.txt"), the unstaged viewer should
show two entries: one for deleting "a.txt" and one for adding "A.txt".

It showed only one.

When populating the viewer content, we must not use case-insensitive
comparisons. Case-insensitive comparisons are fine later for determining
the display order, but if we use case-insensitive comparisons when
creating the entries, the view will have only one entry.

This fixes the display part of [1]. The more severe problem (file
"A.txt" is deleted altogether) will be fixed in JGit.

[1] https://github.com/eclipse-egit/egit/issues/76

Bug: egit-76
Change-Id: Ibf1a3bc5096111298c2bfa0d0a259e38793f7281