StagingView: make use of new JGit AddCommand behavior
JGit's AddCommand newly implements the more modern behavior of C git
and now also stages deletions automatically if paths are given. For
the "Add to index" action in the unstaged viewer it is thus no longer
necessary to split the paths into two lists, one for an AddCommand and
one for a RmCommand. AddCommand can do it all.

Change-Id: If986fb0ef6369c9a80e8d9bf328a3ac69e1a9f99