CommitMessageComponent: improve empty message detection and clean up
I see commits in Gerrit from time to time with a commit message
"Change-Id: I0000000000000000000000000000000000000000". Evidently
the check for an empty commit message fails for an empty message
including the Gerrit change-id footer.

This is caused by cleaning the message according to git configs
commit.cleanup and core.commentChar. An empty commit message in EGit
for a Gerrit-enabled repo is two empty lines followed by the Change-Id
footer. Cleaning up this message removes the two leading empty lines,
after which the footer is no longer a footer but the commit subject.
Thus CommitMessageComponent.checkCommitInfo() no longer considers this
message to be empty, and the user can perform the commit.

(This also happens outside of EGit when one enters such a commit message
in vi in command-line git.)

To avoid at least the common case occurring in EGit, check explicitly
whether the message _before_ cleaning is only empty lines followed by a
Gerrit change-id. If so, treat the message as empty and do not allow to
commit. This is minimally invasive quickfix.

Also do some minor clean-up in that checkICommitInfo() method.

Change-Id: I4b77aef3cb13017866d3abd25442d35d634d1410