CommitMessageComponent: improve empty message detection
Commit 03b84e25 had already handled the special case of a message
containing only the Gerrit Change-Id footer. However, the same problem
can also occur with other footers: if a message contains only a few
blank lines and footers, clean-up will move the footers to the start,
at which point they cease to be footers, and the commit message would
be considered non-empty.

Improve this by checking first for footer lines on the non-cleaned text
and get the first footer line. Then clean the message. If now the first
line of the commit message is that footer line (and wasn't before), this
case occurred. Consider the commit message empty in that case.

Change-Id: Ia64e11d7b38df058d6a0f5db277ec5ceb9370267