Staging View: extensible commit message toolbar
Give the toolbar of the commit message component (the one containing the
"Amend", "Sign Off", "Sign Commit", and "Gerrit Change ID" buttons) an
ID and set it up to include dynamic contributions for that ID.

Do the same in the CommitDialog, replacing the ToolBar and ToolItems
used directly by a ToolBarManager and Actions.

In the staging view, hook up the commit message preview to update if the
commit message changes while the preview is visible.

Provide new API for user code in the form of a CommitContext to get
access to the repository and commit message in commands contributed to
that toolbar.

With this, external code can contribute via plugin.xml its own commands
to the toolbar that can determine the repository state and that can
modify the commit message.

Bug: egit-97
Change-Id: I6bb9552ddc49a13419301ae392700f9e7f4890c2