[themes] Fix light theme in staging and history views
In the staging view, give the staged and unstaged tree viewers a white
background when active. The light gray default background makes them
look inactive, especially since the commit message StyledText behaves
like an editor and has the configured editor background, which is white
by default.

In the history view, remove the gap between the commit message and the
diff part, and also change the background on the commit message to
white. The light gray reduces contrast with the gray text used for the
commit headers too much.

In the dark theme, override this with more specific selectors that set
"background-color: inherit" to get the default dark theme behavior.
The selectors must be more specific because the egit.css is loaded for
*all* themes as it has no themeId in the pom.xml. With selectors of
equal specificity the backgrounds would remain white also in the dark
theme, which don't want of course.

Note that since commit 42fa76d6e, the EGit baseline is Eclipse 4.36
(2025-06), so compatibility of these styling changes with earlier
Eclipse themes is not an issue anymore. EGit 7.4.0 will not be
installable on older Eclipses.

Bug: egit-57
Change-Id: I015ce6e65c750e4f6f647c6d8f836d23d643a77d