BranchOperationsUI: Update UI elements sooner after checkout
Originally, after a checkout was performed, a refresh on the decorator
was performed without updating the decorator's cache. This had the
consequence that the UI (labels, nodes, staging view, etc) were not
really updated since the cache still contained the old state of the
repository i.e. the old branch name.

The code has been corrected and it clears the cache for the affected
repository before asking for a UI refresh, forcing the cache to fetch
the name of the branch again before updating the UI.

Bug: egit-114

Change-Id: I50d7d9d42203d7bbc793fd527337de126bc42462