Reword dialog: fix layout bug hiding short commit messages
Sometimes the layout assigns a negative y-coordinate to the message
area when the dialog is opened on MacOS. It is unknown why exactly
this happens, but in any case layouting again corrects this.

As a work-around force a re-layout the first time the message area
is to be drawn.

Also remove an unneeded Composite, GridData settings that had no effect
since the parent's layout is not a GridLayout, and FormToolkit settings
that also had no effect since the form toolkit isn't used in that
dialog.

Bug: egit-54
Change-Id: I84269dd8d2c7af7b205d10240a36be05b218072d